+++ tests/dummy/config/environment.js
@@ -17,8 +17,8 @@ module.exports = function(environment) {
       // Here you can pass flags/options to your application instance
       // when it is created
     },
-    'ember-moment': {
-      includeTimezone: true
+    moment: {
+      includeTimezone: 'all'
     }
   };